requirements.txt
torch>=2.3.0
transformers>=4.44.0
accelerate>=0.29.0
pillow>=10.3.0
opencv-python>=4.9.0
psutil>=5.9.0
